requirements.txt
Flask
dnspython3